(maint) Added logger, streamlined logging, added setters for global options

+++ lib/in_parallel.rb
@@ -1,11 +1,17 @@
+require_relative 'parallel_logger'
 require_relative 'parallel_enumerable'
+require 'tempfile'
+
 module InParallel
+  include ParallelLogger
+
   class InParallelExecutor
     # How many seconds between outputting to stdout that we are waiting for child processes.
     # 0 or < 0 means no signaling.
-    @@signal_interval = 30
-    @@timeout = 1800
-    @@process_infos = []
+    @@parallel_signal_interval = 30
+    @@parallel_default_timeout = 1800
+
+    @@process_infos            = []
     def self.process_infos
       @@process_infos
     end
@@ -21,44 +27,56 @@ module InParallel
       @@main_pid
     end
 
-    def self.timeout
-      @@timeout
+    def self.parallel_default_timeout
+      @@parallel_default_timeout
     end
 
-    def self.timeout=(value)
-      @@timeout = value
+    def self.parallel_default_timeout=(value)
+      @@parallel_default_timeout = value
     end
 
+    def self.logger
+      @@logger
+    end
+
+    def self.logger=(value)
+      @@logger = value
+    end
+
+    # Runs all methods within the block in parallel and waits for them to complete
+    #
     # Example - will spawn 2 processes, (1 for each method) wait until they both complete, and log STDOUT:
-    # InParallel.run_in_parallel {
-    #   @result_1 = method1
-    #   @result_2 = method2
-    # }
+    #   InParallel.run_in_parallel {
+    #     @result_1 = method1
+    #     @result_2 = method2
+    #   }
     # NOTE: Only supports assigning instance variables within the block, not local variables
-    def self.run_in_parallel(timeout = @@timeout, kill_all_on_error = false, &block)
-      if Process.respond_to?(:fork)
+    def self.run_in_parallel(timeout = @@parallel_default_timeout, kill_all_on_error = false, &block)
+      if fork_supported?
         proxy = BlankBindingParallelProxy.new(block.binding)
         proxy.instance_eval(&block)
         return wait_for_processes(proxy, block.binding, timeout, kill_all_on_error)
       end
-      puts 'Warning: Fork is not supported on this OS, executing block normally'
+      # if fork is not supported
       block.call
     end
 
-    # Example - Will spawn a process in the background to run puppet agent on two agents and return immediately:
-    # Parallel.run_in_background {
-    #   @result_1 = method1
-    #   @result_2 = method2
-    # }
-    # # Do something else here before waiting for the process to complete
+    # Runs all methods within the block in parallel in the background
     #
-    # # Optionally wait for the processes to complete before continuing.
-    # # Otherwise use run_in_background(true) to clean up the process status and output immediately.
-    # wait_for_processes(self)
-    # NOTE: must call get_background_results to allow instance variables in calling object to be set,
-    # otherwise @result_1 will evaluate to "unresolved_parallel_result_0"
+    # Example - Will spawn a process in the background to run puppet agent on two agents and return immediately:
+    #   Parallel.run_in_background {
+    #     @result_1 = method1
+    #     @result_2 = method2
+    #   }
+    #   # Do something else here before waiting for the process to complete
+    #
+    #   # Optionally wait for the processes to complete before continuing.
+    #   # Otherwise use run_in_background(true) to clean up the process status and output immediately.
+    #   wait_for_processes(self)
+    #
+    # NOTE: must call get_background_results to allow instance variables in calling object to be set, otherwise @result_1 will evaluate to "unresolved_parallel_result_0"
     def self.run_in_background(ignore_result = true, &block)
-      if Process.respond_to?(:fork)
+      if fork_supported?
         proxy = BlankBindingParallelProxy.new(block.binding)
         proxy.instance_eval(&block)
 
@@ -71,25 +89,22 @@ module InParallel
         end
         return
       end
-      puts 'Warning: Fork is not supported on this OS, executing block normally'
+      # if fork is not supported
       result = block.call
       return nil if ignore_result
       result
     end
 
-    # Waits for all processes to complete and logs STDOUT and STDERR in chunks from any processes
-    # that were triggered from this Parallel class
-    # @param [Object] proxy - The instance of the proxy class that the method was executed within
-    # (probably only useful when called by run_in_background)
-    # @param [Object] binding - The binding of the block to assign return values to instance variables
-    # (probably only useful when called by run_in_background)
+    # Waits for all processes to complete and logs STDOUT and STDERR in chunks from any processes that were triggered from this Parallel class
+    # @param [Object] proxy - The instance of the proxy class that the method was executed within (probably only useful when called by run_in_background)
+    # @param [Object] binding - The binding of the block to assign return values to instance variables (probably only useful when called by run_in_background)
     # @param [Int] timeout Time in seconds to wait before giving up on a child process
-    # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately -
-    # killing all other forked processes - when one process errors.
+    # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately - killing all other forked processes - when one process errors.
     def self.wait_for_processes(proxy = self, binding = nil, timeout = nil, kill_all_on_error = false)
       raise_error = nil
-      timeout ||= @@timeout
+      timeout ||= @@parallel_default_timeout
       trap(:INT) do
+        # Can't use logger inside of trap
         puts "Warning, recieved interrupt.  Processing child results and exiting."
         kill_child_processes
       end
@@ -100,8 +115,8 @@ module InParallel
       start_time = Time.now
       timer = start_time
       while !@@process_infos.empty? do
-        if @@signal_interval > 0 && Time.now > timer + @@signal_interval
-          puts 'Waiting for child processes.'
+        if @@parallel_signal_interval > 0 && Time.now > timer + @@parallel_signal_interval
+          @@logger.debug 'Waiting for child processes.'
           timer = Time.now
         end
         if Time.now > start_time + timeout
@@ -116,21 +131,23 @@ module InParallel
             # the process completed, get the result and rethrow on error.
             begin
               # Print the STDOUT and STDERR for each process with signals for start and end
-              puts "\n------ Begin output for #{process_info[:method_sym]} - #{process_info[:pid]}\n"
-              puts File.new(process_info[:std_out], 'r').readlines
-              puts "------ Completed output for #{process_info[:method_sym]} - #{process_info[:pid]}\n"
+              @@logger.info "------ Begin output for #{process_info[:method_sym]} - #{process_info[:pid]}"
+              # Content from the other thread will already be pre-pended with log stuff (info, warn, date/time, etc)
+              # So don't use logger, just use puts.
+              puts "  " + File.new(process_info[:std_out], 'r').readlines.join("  ")
+              @@logger.info "------ Completed output for #{process_info[:method_sym]} - #{process_info[:pid]}"
               result = process_info[:result].read
               marshalled_result = (result.nil? || result.empty?) ? result : Marshal.load(result)
+              # Kill all other processes and let them log their stdout before re-raising
+              # if a child process raised an error.
               if marshalled_result.is_a?(Exception)
                 raise_error = marshalled_result.dup
                 kill_child_processes if kill_all_on_error
                 marshalled_result = nil
               end
               results_map[process_info[:index]] = {process_info[:tmp_result] => marshalled_result}
-              File.delete(process_info[:std_out])
-              # Kill all other processes and let them log their stdout before re-raising
-              # if a child process raised an error.
             ensure
+              File.delete(process_info[:std_out]) if File.exists?(process_info[:std_out])
               # close the read end pipe
               process_info[:result].close unless process_info[:result].closed?
               @@process_infos.delete(process_info)
@@ -164,21 +181,22 @@ module InParallel
       ret_val = nil
       # Communicate the return value of the method or block
       read_result, write_result = IO.pipe
-      Dir.mkdir('tmp') unless Dir.exists? 'tmp'
       pid = fork do
+        Dir.mkdir('tmp') unless Dir.exists? 'tmp'
+        stdout_file = File.new("tmp/pp_#{Process.pid}", 'w')
         exit_status = 0
         trap(:INT) do
-          puts("Warning: Interrupt received in child process; exiting #{Process.pid}")
+          # Can't use logger inside of trap
+          puts "Warning: Interrupt received in child process; exiting #{Process.pid}"
           kill_child_processes
           return
         end
-        write_file = File.new("tmp/parallel_process_#{Process.pid}", 'w')
 
         # IO buffer is 64kb, which isn't much... if debug logging is turned on,
         # this can be exceeded before a process completes.
         # Storing output in file rather than using IO.pipe
-        STDOUT.reopen(write_file)
-        STDERR.reopen(write_file)
+        STDOUT.reopen(stdout_file)
+        STDERR.reopen(stdout_file)
 
         begin
           # close subprocess's copy of read_result since it only needs to write
@@ -191,7 +209,7 @@ module InParallel
             begin
               ret_val = ret_val.dup
             rescue StandardError => err
-              puts "Warning: return value from child process #{ret_val} " +
+              @@logger.warn "Warning: return value from child process #{ret_val} " +
                        "could not be transferred to parent process: #{err.message}"
             end
           end
@@ -199,11 +217,11 @@ module InParallel
           begin
             Marshal.dump(ret_val, write_result) unless ret_val.nil?
           rescue StandardError => err
-            puts "Warning: return value from child process #{ret_val} " +
+            @@logger.warn "Warning: return value from child process #{ret_val} " +
                      "could not be transferred to parent process: #{err.message}"
           end
         rescue Exception => err
-          puts "Error in process #{pid}: #{err.message}"
+          @@logger.error "Error in process #{pid}: #{err.message}"
           # Return the error if an error is rescued so we can re-throw in the main process.
           Marshal.dump(err, write_result)
           exit_status = 1
@@ -212,6 +230,8 @@ module InParallel
           exit exit_status
         end
       end
+
+      @@logger.info "Forked process for #{method_sym} - PID = '#{pid}'"
       write_result.close
       # Process.detach returns a thread that will be nil if the process is still running and thr if not.
       # This allows us to check to see if processes have exited without having to call the blocking Process.wait functions.
@@ -220,13 +240,19 @@ module InParallel
       process_info = { :wait_thread => wait_thread,
                        :pid => pid,
                        :method_sym => method_sym,
-                       :std_out => "tmp/parallel_process_#{pid}",
+                       :std_out => "tmp/pp_#{pid}",
                        :result => read_result,
                        :tmp_result => "unresolved_parallel_result_#{@@result_id}",
                        :index => @@process_infos.count }
       @@process_infos.push(process_info)
       @@result_id += 1
       process_info
+    end
+
+    def self.fork_supported?
+      @@supported ||= Process.respond_to?(:fork)
+      @@logger.warn 'Warning: Fork is not supported on this OS, executing block normally' unless @@supported
+      @@supported
     end
 
     def self.kill_child_processes
@@ -276,49 +302,65 @@ module InParallel
       def method_missing(method_sym, *args, &block)
         if InParallelExecutor.main_pid == ::Process.pid
           out = InParallelExecutor._execute_in_parallel("'#{method_sym.to_s}' #{caller_locations[0].to_s}", @object.eval('self')) {send(method_sym, *args, &block)}
-          puts "Forked process for '#{method_sym}' - PID = '#{out[:pid]}'\n"
           out[:tmp_result]
         end
       end
     end
   end
 
-  # Executes each method within a block in a different process
+  InParallelExecutor.logger = @logger
+
+  def parallel_signal_interval
+    InParallelExecutor.parallel_signal_interval
+  end
+
+  def parallel_signal_interval=(value)
+    InParallelExecutor.parallel_signal_interval = value
+  end
+
+  def parallel_default_timeout
+    InParallelExecutor.parallel_default_timeout
+  end
+
+  def parallel_default_timeout=(value)
+    InParallelExecutor.parallel_default_timeout = value
+  end
+
+  # Executes each method within a block in a different process.
+  #
   # Example - Will spawn a process in the background to execute each method
-  # Parallel.run_in_parallel {
-  #   @result_1 = method1
-  #   @result_2 = method2
-  # }
-  # NOTE - Only instance variables can be assigned the return values of the methods within the block.
-  # Local variables will not be assigned any values.
+  #   Parallel.run_in_parallel {
+  #     @result_1 = method1
+  #     @result_2 = method2
+  #   }
+  # NOTE - Only instance variables can be assigned the return values of the methods within the block. Local variables will not be assigned any values.
   # @param [Int] timeout Time in seconds to wait before giving up on a child process
-  # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately -
-  # killing all other forked processes - when one process errors.
+  # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately - killing all other forked processes - when one process errors.
   # @param [Block] block This method will yield to a block of code passed by the caller
   # @return [Array<Result>, Result] the return values of each method within the block
   def run_in_parallel(timeout=nil, kill_all_on_error = false, &block)
-    timeout ||= InParallelExecutor.timeout
+    timeout ||= InParallelExecutor.parallel_default_timeout
     InParallelExecutor.run_in_parallel(timeout, kill_all_on_error, &block)
   end
 
-  # Forks a process for each method within a block and returns immediately
+  # Forks a process for each method within a block and returns immediately.
+  #
   # Example 1 - Will fork a process in the background to execute each method and return immediately:
-  # Parallel.run_in_background {
-  #   @result_1 = method1
-  #   @result_2 = method2
-  # }
+  #   Parallel.run_in_background {
+  #     @result_1 = method1
+  #     @result_2 = method2
+  #   }
   #
   # Example 2 - Will fork a process in the background to execute each method, return immediately, then later
   # wait for the process to complete, printing it's STDOUT and assigning return values to instance variables:
-  # Parallel.run_in_background(false) {
-  #   @result_1 = method1
-  #   @result_2 = method2
-  # }
-  # # Do something else here before waiting for the process to complete
+  #   Parallel.run_in_background(false) {
+  #     @result_1 = method1
+  #     @result_2 = method2
+  #   }
+  #   # Do something else here before waiting for the process to complete
   #
-  # wait_for_processes
-  # NOTE: must call wait_for_processes to allow instance variables within the block to be set,
-  # otherwise results will evaluate to "unresolved_parallel_result_X"
+  #   wait_for_processes
+  # NOTE: must call wait_for_processes to allow instance variables within the block to be set, otherwise results will evaluate to "unresolved_parallel_result_X"
   # @param [Boolean] ignore_result True if you do not care about the STDOUT or return value of the methods executing in the background
   # @param [Block] block This method will yield to a block of code passed by the caller
   # @return [Array<Result>, Result] the return values of each method within the block
@@ -326,14 +368,12 @@ module InParallel
     InParallelExecutor.run_in_background(ignore_result, &block)
   end
 
-  # Waits for all processes started by run_in_background to complete execution, then prints STDOUT
-  # and assigns return values to instance variables.  See :run_in_background
+  # Waits for all processes started by run_in_background to complete execution, then prints STDOUT and assigns return values to instance variables.  See :run_in_background
   # @param [Int] timeout Time in seconds to wait before giving up on a child process
-  # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately -
-  # killing all other forked processes - when one process errors.
+  # @param [Boolean] kill_all_on_error Whether to wait for all processes to complete, or fail immediately - killing all other forked processes - when one process errors.
   # @return [Array<Result>, Result] the temporary return values of each method within the block
   def wait_for_processes(timeout=nil, kill_all_on_error = false)
-    timeout ||= InParallelExecutor.timeout
+    timeout ||= InParallelExecutor.parallel_default_timeout
     InParallelExecutor.wait_for_processes(nil, nil, timeout, kill_all_on_error)
   end
 end
